refactor: remove MonsterStore dependency from Monster, remove MonsterStore
