+++ public/examples/shoot-vr/index.html
@@ -172,7 +172,10 @@
       // XR controllers
       const raycaster = new THREE.Raycaster();
 
-      function onSelectStart() {
+      function onSelectStart(event) {
+        const symbols = Object.getOwnPropertySymbols(event.data);
+        console.log(event.data[symbols[0]].impl.handedness);
+        console.log(event);
         shootAudio.stop();
         shootAudio.play();
